new changes in hm10
